hz - fixed most of the UI but toolbars need to be consistent, also google map, and scroll bar pending
