+++ app/src/main/java/com/ucsb/integration/Login/BasicLoginActivity.java
@@ -2,7 +2,6 @@ package com.ucsb.integration.Login;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.text.TextUtils;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -15,47 +14,68 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
 import com.ucsb.integration.MainActivity;
 import com.ucsb.integration.R;
 
-import java.util.HashMap;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.net.wifi.hotspot2.pps.Credential;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+import android.widget.Button;
+import android.widget.Toast;
+
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.common.SignInButton;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.common.api.ApiException;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthCredential;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.auth.GoogleAuthProvider;
+
+
+
 
 public class BasicLoginActivity extends AppCompatActivity {
 
+    private SignInButton signInButton;
+    private static final int RC_SIGN_IN=9001;
+    private GoogleSignInClient mGoogleSignInClient;
+    private String TAG ="BasicLoginActivity";
+    private FirebaseAuth mAuth;
+    //private Button btnSignOut;
     EditText email;
     EditText password;
     Button signup;
     Button login;
-    Button gsign_in_button;
-    Button fbsign_in_button;
-    FirebaseAuth firebaseAuth;
-    DatabaseReference reference;
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_basic_login);
-
+        setContentView(R.layout.content_basic_login);
 
         email=findViewById(R.id.email);
         password=findViewById(R.id.password);
         signup=findViewById(R.id.signup);
         login=findViewById(R.id.login);
-        gsign_in_button=findViewById(R.id.gsign_in_button);
-        fbsign_in_button=findViewById(R.id.fbsign_in_button);
 
+        signInButton = (SignInButton) findViewById(R.id.gsign_in_button);
+        mAuth =FirebaseAuth.getInstance();
+        //btnSignOut= findViewById(R.id.sign_out_button);
 
-
-        firebaseAuth=FirebaseAuth.getInstance();
-
-        /*signup.setOnClickListener(new View.OnClickListener() {
+        signup.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                firebaseAuth.createUserWithEmailAndPassword(email.getText().toString(),password.getText().toString()).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                mAuth.createUserWithEmailAndPassword(email.getText().toString(),password.getText().toString()).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
                     @Override
                     public void onComplete(@NonNull Task<AuthResult> task) {
                         if(task.isSuccessful()){
@@ -68,26 +88,12 @@ public class BasicLoginActivity extends AppCompatActivity {
                     }
                 });
             }
-        });*/
-
-        signup.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String txt_email = email.getText().toString();
-                String txt_password = password.getText().toString();
-
-                if (TextUtils.isEmpty(txt_email) || TextUtils.isEmpty(txt_password)) {
-                    Toast.makeText(BasicLoginActivity.this, "All fields are required", Toast.LENGTH_SHORT).show();
-                } else {
-                    register(txt_email, txt_password);
-                }
-            }
         });
 
         login.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                firebaseAuth.signInWithEmailAndPassword(email.getText().toString(),password.getText().toString()).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                mAuth.signInWithEmailAndPassword(email.getText().toString(),password.getText().toString()).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
                     @Override
                     public void onComplete(@NonNull Task<AuthResult> task) {
                         if(task.isSuccessful()){
@@ -102,55 +108,89 @@ public class BasicLoginActivity extends AppCompatActivity {
             }
         });
 
-        gsign_in_button.setOnClickListener(new View.OnClickListener() {
+
+
+        // Configure Google Sign In
+        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestIdToken(getString(R.string.default_web_client_id))
+                .requestEmail()
+                .build();
+
+        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
+
+        signInButton.setOnClickListener(new View.OnClickListener(){
             @Override
-            public void onClick(View v) {
-                startActivity(new Intent(BasicLoginActivity.this, GoogleLoginActivity.class));
+            public void onClick(View v){
+                signIn();
             }
         });
 
-        fbsign_in_button.setOnClickListener(new View.OnClickListener() {
+       /* btnSignOut.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                startActivity(new Intent(BasicLoginActivity.this, FacebookLoginActivity.class));
+                mGoogleSignInClient.signOut();
+                Toast.makeText(BasicLoginActivity.this, "You are Logged Out", Toast.LENGTH_SHORT).show();
+                btnSignOut.setVisibility(View.INVISIBLE);
+                //Should be Invisible at the beginning(Unsolved)
             }
         });
+*/
+
+    }
+    private void signIn() {
+        Intent signInIntent = mGoogleSignInClient.getSignInIntent();
+        startActivityForResult(signInIntent, RC_SIGN_IN);
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        if(requestCode==RC_SIGN_IN){
+            Task<GoogleSignInAccount> task=GoogleSignIn.getSignedInAccountFromIntent(data);
+            handleSignInResult(task);
+        }
 
     }
 
-    private void register(final String email, final String password) {
-        firebaseAuth.createUserWithEmailAndPassword(email, password)
-                .addOnCompleteListener(BasicLoginActivity.this,new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()) {
-                            FirebaseUser firebaseUser = firebaseAuth.getCurrentUser();
-                            assert firebaseUser != null;
-                            String userid = firebaseUser.getUid();
+    private void handleSignInResult (Task<GoogleSignInAccount> completeTask){
+        try{
+            GoogleSignInAccount acc= completeTask.getResult(ApiException.class);
+            Toast.makeText(BasicLoginActivity.this, "Signed In Successfully", Toast.LENGTH_SHORT).show();
 
-                            reference = FirebaseDatabase.getInstance().getReference("Users").child(userid);
+            firebaseAuthWithGoogle(acc);
+            startActivity(new Intent(BasicLoginActivity.this, MainActivity.class));
 
-                            HashMap<String, String> hashMap = new HashMap<>();
-                            hashMap.put("id", userid);
-                            hashMap.put("email", email);
-                            hashMap.put("imageURL", "default");
-
-                            reference.setValue(hashMap).addOnCompleteListener(new OnCompleteListener<Void>() {
-                                @Override
-                                public void onComplete(@NonNull Task<Void> task) {
-                                    if (task.isSuccessful()) {
-                                        Intent intent = new Intent(BasicLoginActivity.this, MainActivity.class);
-                                        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
-                                        startActivity(intent);
-                                        finish();
-                                    }
-                                }
-                            });
-                        } else {
-                            Toast.makeText(BasicLoginActivity.this, "You can't register with this email or password", Toast.LENGTH_SHORT).show();
-                        }
-                    }
-                });
+        }
+        catch(ApiException e){
+            Toast.makeText(BasicLoginActivity.this, "Failed to Signed In", Toast.LENGTH_SHORT).show();
+            firebaseAuthWithGoogle(null);
+        }
     }
+
+    private void firebaseAuthWithGoogle(GoogleSignInAccount acc) {
+        AuthCredential credential = GoogleAuthProvider.getCredential(acc.getIdToken(), null);
+        mAuth.signInWithCredential(credential).addOnCompleteListener(this, new OnCompleteListener<AuthResult>(){
+            @Override
+            public void onComplete(@NonNull Task<AuthResult> task) {
+                if (task.isSuccessful()) {
+                    // Sign in success, update UI with the signed-in user's information
+                    Log.d(TAG, "signInWithCredential:success");
+                    FirebaseUser user = mAuth.getCurrentUser();
+                    updateUI(user);
+                } else {
+                    // If sign in fails, display a message to the user.
+                    Log.w(TAG, "signInWithCredential:failure", task.getException());
+                    updateUI(null);
+                }
+            }
+        });
+    }
+
+    //updateUI for Later use
+    private void updateUI(FirebaseUser fuser){
+        //btnSignOut.setVisibility(View.VISIBLE);
+        Toast.makeText(BasicLoginActivity.this, "Success", Toast.LENGTH_SHORT).show();
+    }
+
 
 }
