+++ app/src/main/java/com/ucsb/integration/MainActivity.java
@@ -91,5 +91,6 @@ public class MainActivity extends AppCompatActivity {
         });
     }
 
+    //getSupportActionBar().hide();
 
 }
