+++ app/src/main/res/layout/activity_start.xml
@@ -27,5 +27,6 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.427" />
+        app:layout_constraintVertical_bias="0.427"
+        android:textColor="#000000" />
 </androidx.constraintlayout.widget.ConstraintLayout>
